+++ src/dev-tools.ts
@@ -1,4 +1,5 @@
 import * as _ from "lodash";
+import {isTyduxDevelopmentModeEnabled} from "./development";
 import {getGlobalTyduxState, globalStateChanges$} from "./global-state";
 
 interface DevToolsState {
@@ -47,9 +48,10 @@ export function enableDevTools() {
             .subscribe(event => {
                 const mutator = event.boundMutator ? event.boundMutator : _.noop;
                 mutators.push(mutator);
+                const meta = event.duration !== undefined ? ` (${event.duration}ms)` : "";
                 const action = {
                     ...event.action,
-                    "type": "[" + event.storeName + " / " + event.action.type + "]"
+                    "type": "[" + event.storeName + " / " + event.action.type + "]" + meta
                 };
 
                 devTools.send(action, getGlobalTyduxState());